Overwrite protection active by default
Added an overwrite protection to LocalStorage and S3Storage. This will
prevent accidental overwrites of data when trying to `write` a path that
already exists. The normal flow would be first delete (on purpose) and
then write again if the user wants to do that.

The behaviour can be switched off, but it's a very sane default.