FIX: Move main module to an inner name
Asking the clients to include Saviour top level module was a bad idea
definetly. This caused an issue for including classes that were using
the default ruby `File` constant, because Saviour defines
`Saviour::File` and this constant was taking precendence due to the
inclusion. Now the main saviour funcionality is provided by
`Saviour::Model` instead of `Saviour` directly.

Also fixed a minor out of date syntaxes in the readme.