Avoid extra update with only updated_at
When saving a file that has an attachment and only saving the
attachment (without updating any other attribute(, there should
be first an empty update that causes no sql query (because rails
detects nothing has changed at the db level, even if we do will_change!
on the virtual attribute).

However this is not the case, and rails issues an update call
that only updates `updated_at`.

Under the covers, active record trusts dirty attribute tracking
to know if something's changed or not, so it's not possible to change
this behaviour while keep using AM dirty tracking (will_change!).

This commit replaces that usage for a custom path on all the public
interface of dirty tracking, so it becomes transparent for the user,
but without using AM::Dirty.